+++ griptape/events/event_listener.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Callable, Optional
 
-from attrs import Factory, define, field
+from attrs import define, field
 
 if TYPE_CHECKING:
     from griptape.drivers import BaseEventListenerDriver
@@ -12,9 +12,20 @@ if TYPE_CHECKING:
 
 @define
 class EventListener:
-    handler: Callable[[BaseEvent], Optional[dict]] = field(default=Factory(lambda: lambda event: event.to_dict()))
+    """An event listener that listens for events and handles them.
+
+    Args:
+        handler: The handler function that will be called when an event is published.
+            The handler function should accept an event and return either the event or a dictionary.
+            If the handler returns None, the event will not be published.
+        event_types: A list of event types that the event listener should listen for.
+            If not provided, the event listener will listen for all event types.
+        event_listener_driver: The driver that will be used to publish events.
+    """
+
+    handler: Optional[Callable[[BaseEvent], Optional[BaseEvent | dict]]] = field(default=None)
     event_types: Optional[list[type[BaseEvent]]] = field(default=None, kw_only=True)
-    driver: Optional[BaseEventListenerDriver] = field(default=None, kw_only=True)
+    event_listener_driver: Optional[BaseEventListenerDriver] = field(default=None, kw_only=True)
 
     _last_event_listeners: Optional[list[EventListener]] = field(default=None)
 
@@ -36,12 +47,12 @@ class EventListener:
         event_types = self.event_types
 
         if event_types is None or type(event) in event_types:
-            event_payload = self.handler(event)
-            if self.driver is not None:
-                if event_payload is not None and isinstance(event_payload, dict):
-                    self.driver.publish_event(event_payload)
-                else:
-                    self.driver.publish_event(event)
+            handled_event = event
+            if self.handler is not None:
+                handled_event = self.handler(event)
 
-        if self.driver is not None and flush:
-            self.driver.flush_events()
+            if self.event_listener_driver is not None and handled_event is not None:
+                self.event_listener_driver.publish_event(handled_event)
+
+        if self.event_listener_driver is not None and flush:
+            self.event_listener_driver.flush_events()